+++ frontend/packages/volto-light-theme/src/customizations/volto/components/manage/Blocks/Block/Edit.jsx
@@ -3,6 +3,8 @@
  * REASON: Adding BlockModelv3 wrappers
  * DATE: 2024-02-01
  * DEVELOPER: @sneridagh
+ * FILE: https://github.com/plone/volto/blob/19.3.0/packages/volto/src/components/manage/Blocks/Block/Edit.jsx
+ * FILE VERSION: Volto 19.3.0
  */
 
 import Edit from '../../../../../../components/Blocks/Block/Edit';