+++ frontend/packages/volto-light-theme/src/customizations/volto/components/manage/Blocks/Image/Edit.jsx
@@ -2,6 +2,9 @@
  * OVERRIDE Edit.jsx
  * REASON: This theme introduces the concept of "Caption" and uses `figure`
  * tag for the image.
+ * FILE: https://github.com/plone/volto/blob/19.3.0/packages/volto/src/components/manage/Blocks/Image/Edit.jsx
+ * FILE VERSION: Volto 19.3.0
+ * DATE: 2023-07-28
  */
 
 import ImageEdit from '../../../../../../components/Blocks/Image/Edit';