+++ frontend/packages/volto-light-theme/src/customizations/volto/components/manage/Blocks/Image/ImageSidebar.jsx
@@ -2,6 +2,9 @@
  * OVERRIDE ImageSidebar.jsx
  * REASON: There are some specific behavior that we want to enforce to the sizes
  * and alignment settings
+ * FILE: https://github.com/plone/volto/blob/19.3.0/packages/volto/src/components/manage/Blocks/Image/ImageSidebar.jsx
+ * FILE VERSION: Volto 19.3.0
+ * DATE: 2023-07-28
  */
 
 import ImageSidebar from '../../../../../../components/Blocks/Image/ImageSidebar';