+++ frontend/packages/volto-light-theme/src/customizations/volto/components/manage/Blocks/Maps/View.jsx
@@ -3,6 +3,8 @@
  * REASON: Use align:noprefix and blockWidth:noprefix CSS custom properties instead of align class.
  * DATE: 2026-04-16
  * DEVELOPER: @danalvrz
+ * FILE: https://github.com/plone/volto/blob/19.3.0/packages/volto/src/components/manage/Blocks/Maps/View.jsx
+ * FILE VERSION: Volto 19.3.0
  */
 
 import MapsView from '../../../../../../components/Blocks/Maps/View';