+++ frontend/packages/volto-light-theme/src/customizations/volto/components/manage/Blocks/Search/components/SearchInput.jsx
@@ -1,6 +1,9 @@
 /**
  * OVERRIDE SearchInput.jsx
  * REASON: Special template for displaying this component
+ * FILE: https://github.com/plone/volto/blob/19.3.0/packages/volto/src/components/manage/Blocks/Search/components/SearchInput.jsx
+ * FILE VERSION: Volto 19.3.0
+ * DATE: 2023-01-19
  */
 
 import SearchInput from '../../../../../../../components/Blocks/Search/components/SearchInput';