+++ frontend/packages/volto-light-theme/src/customizations/volto/components/theme/Anontools/Anontools.jsx
@@ -4,6 +4,9 @@
  * SemanticUI-free located at the components folder.
  * To override it, override the @kitconcept/volto-light-theme one instead of
  * this one.
+ * FILE: https://github.com/plone/volto/blob/19.3.0/packages/volto/src/components/theme/Anontools/Anontools.jsx
+ * FILE VERSION: Volto 19.3.0
+ * DATE: 2022-11-03
  */
 
 import Anontools from '../../../../../components/Anontools/Anontools';