+++ frontend/packages/volto-light-theme/src/customizations/volto/components/theme/View/NewsItemView.jsx
@@ -3,6 +3,8 @@
  * REASON: Adding dates and headtitle above the title of page.
  * DATE: 2023-07-04
  * DEVELOPER: @IFlameing
+ * FILE: https://github.com/plone/volto/blob/19.3.0/packages/volto/src/components/theme/View/NewsItemView.jsx
+ * FILE VERSION: Volto 19.3.0
  */
 
 import NewsItemView from '../../../../../components/Theme/NewsItemView';